Remove dependency of operations on command-line arguments.
- Move argument parsing into CLI scripts.
- Clean up doc comments for scripts.
- Remove stray comments rendered obsolete by refactoring.
