+++ src/vtp/cli/cast_ballot.py
@@ -21,9 +21,67 @@
 
 See 'cast_ballot.py -h' for usage information.
 """
+
+# Standard imports
+import argparse
 import sys
 
+# Local imports
+from vtp.core.address import Address
+from vtp.core.common import Common
 from vtp.ops.cast_ballot_operation import CastBallotOperation
+
+
+def parse_arguments(argv):
+    """
+    Parse command line arguments.  This can be called either with
+    sys.argv sans the first arg which is the 'script name', in
+    which case argv is a list of strings, or with a dictionary, in
+    which case is converted to a list of strings.
+
+    So to match native argparse argv parsing, if a dict value is a
+    boolean, the value is removed from the list and the key is
+    either kept (True) or deleted (False).  If the value is None,
+    the key is removed.
+    """
+
+    safe_args = Common.cast_thing_to_list(argv)
+    parser = argparse.ArgumentParser(
+        formatter_class=argparse.RawDescriptionHelpFormatter,
+        description="""
+Given either an address or a specific blank ballot, will either read a
+blank ballot and allow a user to manually select choices or when in
+demo mode, cast_ballot.py will randominly select choices.
+""",
+    )
+
+    Address.add_address_args(parser)
+    # ZZZ - cloaked contests are enabled at cast_ballot time
+    #    parser.add_argument('-k', "--cloak", action="store_true",
+    #                            help="if possible provide a cloaked ballot offset")
+    parser.add_argument(
+        "--demo_mode",
+        action="store_true",
+        help="set demo mode to automatically cast random ballots",
+    )
+    parser.add_argument(
+        "--blank_ballot",
+        help="overrides an address - specifies the specific blank ballot",
+    )
+    parser.add_argument(
+        "-v",
+        "--verbosity",
+        type=int,
+        default=3,
+        help="0 critical, 1 error, 2 warning, 3 info, 4 debug (def=3)",
+    )
+    parser.add_argument(
+        "-n",
+        "--printonly",
+        action="store_true",
+        help="will printonly and not write to disk (def=True)",
+    )
+    return parser.parse_args(safe_args)
 
 
 # pylint: disable=duplicate-code
@@ -36,9 +94,9 @@ def main():
     source file.
     """
 
-    # do it
-    cbo = CastBallotOperation(sys.argv[1:])
-    cbo.run()
+    args = parse_arguments(sys.argv[1:])
+    op = CastBallotOperation(args)
+    op.run()
 
 
 # If called directly via this file
