+++ src/vtp/cli/create_blank_ballot.py
@@ -26,15 +26,50 @@ See ../../docs/tech/executable-overview.md for the context in which this file wa
 """
 
 # Standard imports
+import argparse
 import sys
 
 # Local imports
+from vtp.core.address import Address
+from vtp.core.common import Common
 from vtp.ops.create_blank_ballot_operation import CreateBlankBallotOperation
 
 
-################
-# main
-################
+def parse_arguments(argv):
+    """Parse arguments from a command line or from the constructor"""
+
+    safe_args = Common.cast_thing_to_list(argv)
+    parser = argparse.ArgumentParser(
+        formatter_class=argparse.RawDescriptionHelpFormatter,
+        description="""
+Will parse all the config and address_map yaml files in the current
+VTP ElectionData git tree and create a blank ballot based on the
+supplied address.
+""",
+    )
+    Address.add_address_args(parser)
+    parser.add_argument(
+        "-l",
+        "--language",
+        default="en",
+        help="will print the ballot in the specified language",
+    )
+    parser.add_argument(
+        "-v",
+        "--verbosity",
+        type=int,
+        default=3,
+        help="0 critical, 1 error, 2 warning, 3 info, 4 debug (def=3)",
+    )
+    parser.add_argument(
+        "-n",
+        "--printonly",
+        action="store_true",
+        help="will printonly and not write to disk (def=True)",
+    )
+    return parser.parse_args(safe_args)
+
+
 def main():
     """
     Called via a python local install entrypoint or by running this
@@ -44,13 +79,11 @@ def main():
     the source file.
     """
 
-    # do it
-    cbbo = CreateBlankBallotOperation(sys.argv[1:])
-    cbbo.run()
+    args = parse_arguments(sys.argv[1:])
+    op = CreateBlankBallotOperation(args)
+    op.run()
 
 
 # If called directly via this file
 if __name__ == "__main__":
     main()
-
-# EOF
