+++ src/vtp/cli/verify_ballot_receipt.py
@@ -24,10 +24,79 @@ See 'verify_ballot_receipt.py -h' for usage information.
 """
 
 # Standard imports
+import argparse
 import sys
 
 # Local imports
+from vtp.core.address import Address
+from vtp.core.common import Common
 from vtp.ops.verify_ballot_receipt_operation import VerifyBallotReceiptOperation
+
+
+def parse_arguments(argv):
+    """Parse arguments from a command line or from the constructor"""
+
+    safe_args = Common.cast_thing_to_list(argv)
+    parser = argparse.ArgumentParser(
+        formatter_class=argparse.RawDescriptionHelpFormatter,
+        description="""
+Will read a voter's ballot receipt and validate all the digests
+contained therein.  If a contest has been merged to the main branch,
+will report the current ballot tally number (which ballot in the
+actula tally cound is the voter's).
+
+An address is also supported as an argument in which case the last
+ballot check is read from the default location for the specified
+address.
+
+Can also optionally print the ballot's CVRs when a specific ballot
+check row is provided.
+""",
+    )
+
+    Address.add_address_args(parser, True)
+    parser.add_argument(
+        "-f",
+        "--receipt_file",
+        default="",
+        help="specify the ballot receipt location - overrides an address",
+    )
+    parser.add_argument(
+        "-r",
+        "--row",
+        default="",
+        help="specify a row to inspect that row (the first row is 1, not 0)",
+    )
+    parser.add_argument(
+        "-c",
+        "--cvr",
+        action="store_true",
+        help="display the contents of the content CVRs specifying a row",
+    )
+    parser.add_argument(
+        "-x",
+        "--do_not_pull",
+        action="store_true",
+        help="Before tallying the votes, pull the ElectionData repo",
+    )
+    parser.add_argument(
+        "-v",
+        "--verbosity",
+        type=int,
+        default=3,
+        help="0 critical, 1 error, 2 warning, 3 info, 4 debug (def=3)",
+    )
+    #    parser.add_argument("-n", "--printonly", action="store_true",
+    #                            help="will printonly and not write to disk (def=True)")
+
+    parsed_args = parser.parse_args(safe_args)
+
+    # Validate required args
+    if not (parsed_args.receipt_file or (parsed_args.state and parsed_args.town)):
+        raise ValueError(
+            "Either an explicit or implicit (via an address) receipt file must be provided"
+        )
+    return parsed_args
 
 
 def main():
@@ -39,9 +108,9 @@ def main():
     in the source file.
     """
 
-    # do it
-    vbro = VerifyBallotReceiptOperation(sys.argv[1:])
-    vbro.run()
+    args = parse_arguments(sys.argv[1:])
+    op = VerifyBallotReceiptOperation(args)
+    op.run()
 
 
 # If called directly via this file
