+++ src/vtp/ops/generate_all_blank_ballots_operation.py
@@ -23,7 +23,6 @@ See 'generate_all_blank_ballots.py -h' for usage information.
 """
 
 # Standard imports
-import argparse
 import logging
 import os
 import pprint
@@ -43,47 +42,15 @@ class GenerateAllBlankBallotsOperation:
     description (immediately below this) in the source file.
     """
 
-    @staticmethod
-    def parse_arguments(argv):
-        """Parse arguments from a command line or from the constructor"""
-
-        safe_args = Common.cast_thing_to_list(argv)
-        parser = argparse.ArgumentParser(
-            formatter_class=argparse.RawDescriptionHelpFormatter,
-            description="""
-    Will crawl the ElectionData tree and determine all possible blank
-    ballots and generate them.  They will be placed in the town's
-    blank-ballots subdir.
-    """,
-        )
-
-        parser.add_argument(
-            "-v",
-            "--verbosity",
-            type=int,
-            default=3,
-            help="0 critical, 1 error, 2 warning, 3 info, 4 debug (def=3)",
-        )
-        parser.add_argument(
-            "-n",
-            "--printonly",
-            action="store_true",
-            help="will printonly and not write to disk (def=True)",
-        )
-
-        return parser.parse_args(safe_args)
-
-    def __init__(self, unparsed_args):
+    def __init__(self, args):
         """Only to module-ize the scripts and keep things simple and idiomatic."""
-        self.parsed_args = GenerateAllBlankBallotsOperation.parse_arguments(
-            unparsed_args
-        )
+        self.args = args
 
     def run(self):
         """Main function - see -h for more info"""
 
         # Configure logging
-        Common.configure_logging(self.parsed_args.verbosity)
+        Common.configure_logging(self.args.verbosity)
 
         # Create a VTP ElectionData object if one does not already exist
         the_election_config = ElectionConfig.configure_election()
@@ -118,7 +85,7 @@ class GenerateAllBlankBallotsOperation:
                         pprint.pformat(generic_ballot.dict()),
                     )
                     # Write it out
-                    if self.parsed_args.printonly:
+                    if self.args.printonly:
                         ballot_file = the_election_config.gen_blank_ballot_location(
                             generic_address.active_ggos,
                             generic_address.ballot_subdir,
@@ -129,8 +96,3 @@ class GenerateAllBlankBallotsOperation:
                             the_election_config
                         )
                     logging.info("Blank ballot file: %s", ballot_file)
-
-
-# End Of Class
-
-# EOF
