+++ src/vtp/ops/run_mock_election_operation.py
@@ -25,7 +25,6 @@ See 'run_mock_election.py -h' for usage information.
 """
 
 # Standard imports
-import argparse
 import logging
 import os
 import time
@@ -50,119 +49,9 @@ class RunMockElectionOperation:
     description (immediately below this) in the source file.
     """
 
-    @staticmethod
-    def parse_arguments(argv):
-        """Parse arguments from a command line or from the constructor"""
-
-        safe_args = Common.cast_thing_to_list(argv)
-        parser = argparse.ArgumentParser(
-            formatter_class=argparse.RawDescriptionHelpFormatter,
-            description="""
-Will run a mock election with N ballots across the available blank
-ballots found in the ElectionData.
-
-One basic idea is to run this in different windows, one per VTP
-scanner.  The scanner is nominally associated with a town (as
-configured).
-
-When "-d scanner" is supplied, run_mock_election.py will randomly
-cast and scan ballots.
-
-When "-d server" is supplied, run_mock_election.py will
-synchronously run the merge_contests.py program which will once
-every 10 seconds.  Note that nominally 100 contgests need to have
-been pushed for merge_contests.py to merge in a contest into the
-main branch without the --flush_mode option.
-
-If "-d both" is supplied, run_mock_election.py will run a single
-scanner N iterations while also calling the server function.  If
---flush_mode is set to 1 or 2, run_mock_election.py will then
-flush the ballot cache before printing the tallies and exiting.
-
-By default run_mock_election.py will loop over all available blank
-ballots found withint the ElectionData tree.  However, either a
-specific blank ballot or an address can be specified to limit the
-mock to a single ballot N times.
-""",
-        )
-
-        Address.add_address_args(parser)
-        parser.add_argument(
-            "--blank_ballot",
-            help="overrides an address - specifies the specific blank ballot",
-        )
-        parser.add_argument(
-            "-d",
-            "--device",
-            default="",
-            help="specify a specific VC local device (scanner or server or both) to mock",
-        )
-        parser.add_argument(
-            "-m",
-            "--minimum_cast_cache",
-            type=int,
-            default=100,
-            help="the minimum number of cast ballots required prior to merging (def=100)",
-        )
-        # Note - the black formatter will by default break the help
-        # line below into two lines via a '()', which breaks the
-        # parser.add_argument.  It is python.
-        #
-        # fmt: off
-        parser.add_argument(
-            "-f",
-            "--flush_mode",
-            type=int,
-            default=0,
-            help="will either not flush (0 - default), flush on exit (1), or flush on each iteration (2)",
-        )
-        # fmt: on
-        parser.add_argument(
-            "-i",
-            "--iterations",
-            type=int,
-            default=10,
-            help="the number of unique blank ballots for the scanner app to cast (def=10)",
-        )
-        parser.add_argument(
-            "-u",
-            "--duration",
-            type=int,
-            default=10,
-            help="the number of minutes for the server app to run (def=10)",
-        )
-        parser.add_argument(
-            "-v",
-            "--verbosity",
-            type=int,
-            default=3,
-            help="0 critical, 1 error, 2 warning, 3 info, 4 debug (def=3)",
-        )
-        parser.add_argument(
-            "-n",
-            "--printonly",
-            action="store_true",
-            help="will printonly and not write to disk (def=True)",
-        )
-
-        parsed_args = parser.parse_args(safe_args)
-
-        # Validate required args
-        if parsed_args.device not in ["scanner", "server", "both"]:
-            raise ValueError(
-                "The --device parameter only accepts 'device' or 'server' "
-                f"or 'both' - ({parsed_args.device}) was suppllied."
-            )
-        if parsed_args.flush_mode not in [0, 1, 2]:
-            raise ValueError(
-                "The value of flush_mode must be either 0, 1, or 2"
-                f" - {parsed_args.flush_mode} was supplied."
-            )
-        return parsed_args
-
-    def __init__(self, unparsed_args):
+    def __init__(self, args):
         """Only to module-ize the scripts and keep things simple and idiomatic."""
-        self.parsed_args = RunMockElectionOperation.parse_arguments(unparsed_args)
+        self.args = args
 
     def scanner_mockup(self, the_election_config, ballot):
         """Simulate a VTP scanner"""
@@ -190,12 +79,12 @@ mock to a single ballot N times.
         # Loop over the list N times
         if not blank_ballots:
             raise ValueError("found no blank ballots to cast")
-        for count in range(self.parsed_args.iterations):
+        for count in range(self.args.iterations):
             for blank_ballot in blank_ballots:
                 logging.debug(
                     "Iteration %s of %s - processing %s",
                     count,
-                    self.parsed_args.iterations,
+                    self.args.iterations,
                     blank_ballot,
                 )
                 # - cast a ballot
@@ -203,8 +92,8 @@ mock to a single ballot N times.
                 with Shellout.changed_cwd(election_data_dir):
                     Shellout.run(
                         ["git", "pull"],
-                        printonly=self.parsed_args.printonly,
-                        verbosity=self.parsed_args.verbosity,
+                        printonly=self.args.printonly,
+                        verbosity=self.args.verbosity,
                         no_touch_stds=True,
                         timeout=None,
                         check=True,
@@ -213,8 +102,8 @@ mock to a single ballot N times.
                     {
                         "blank_ballot": blank_ballot,
                         "demo_mode": True,
-                        "verbosity": self.parsed_args.verbosity,
-                        "printonly": self.parsed_args.printonly,
+                        "verbosity": self.args.verbosity,
+                        "printonly": self.args.printonly,
                     }
                 )
                 cast_ballot.run()
@@ -222,21 +111,21 @@ mock to a single ballot N times.
                 accept_ballot = AcceptBallotOperation(
                     {
                         "cast_ballot": Ballot.get_cast_from_blank(blank_ballot),
-                        "verbosity": self.parsed_args.verbosity,
-                        "printonly": self.parsed_args.printonly,
+                        "verbosity": self.args.verbosity,
+                        "printonly": self.args.printonly,
                     }
                 )
                 accept_ballot.run()
-                if self.parsed_args.device == "both":
+                if self.args.device == "both":
                     # - merge the ballot's contests
-                    if self.parsed_args.flush_mode == 2:
+                    if self.args.flush_mode == 2:
                         # Since casting and merging is basically
                         # synchronous, no need for an extra large timeout
                         merge_contests = MergeContestsOperation(
                             {
                                 "flush_mode": True,
-                                "verbosity": self.parsed_args.verbosity,
-                                "printonly": self.parsed_args.printonly,
+                                "verbosity": self.args.verbosity,
+                                "printonly": self.args.printonly,
                             }
                         )
                         merge_contests.run()
@@ -246,9 +135,9 @@ mock to a single ballot N times.
                         # timeout
                         merge_contests = MergeContestsOperation(
                             {
-                                "minimum_cast_cache": self.parsed_args.minimum_cast_cache,
-                                "verbosity": self.parsed_args.verbosity,
-                                "printonly": self.parsed_args.printonly,
+                                "minimum_cast_cache": self.args.minimum_cast_cache,
+                                "verbosity": self.args.verbosity,
+                                "printonly": self.args.printonly,
                             }
                         )
                         merge_contests.run()
@@ -256,36 +145,36 @@ mock to a single ballot N times.
                     if count % 10 == 9:
                         Shellout.run(
                             ["git", "gc"],
-                            printonly=self.parsed_args.printonly,
-                            verbosity=self.parsed_args.verbosity,
+                            printonly=self.args.printonly,
+                            verbosity=self.args.verbosity,
                             no_touch_stds=True,
                             timeout=None,
                             check=True,
                         )
-        if self.parsed_args.device == "both":
+        if self.args.device == "both":
             # merge the remaining contests
             # Note - this needs a longer timeout as it can take many seconds
             merge_contests = MergeContestsOperation(
                 {
                     "flush_mode": True,
-                    "verbosity": self.parsed_args.verbosity,
-                    "printonly": self.parsed_args.printonly,
+                    "verbosity": self.args.verbosity,
+                    "printonly": self.args.printonly,
                 }
             )
             merge_contests.run()
             # tally the contests
             tally_contests = TallyContestsOperation(
                 {
-                    "verbosity": self.parsed_args.verbosity,
-                    "printonly": self.parsed_args.printonly,
+                    "verbosity": self.args.verbosity,
+                    "printonly": self.args.printonly,
                 }
             )
             tally_contests.run()
         # clean up git just in case
         Shellout.run(
             ["git", "gc"],
-            printonly=self.parsed_args.printonly,
-            verbosity=self.parsed_args.verbosity,
+            printonly=self.args.printonly,
+            verbosity=self.args.verbosity,
             no_touch_stds=True,
             timeout=None,
             check=True,
@@ -299,7 +188,7 @@ mock to a single ballot N times.
         # the branches to be merged are remote and not local.
         start_time = time.time()
         # Loop for a day and sleep for 10 seconds
-        seconds = 60 * self.parsed_args.duration
+        seconds = 60 * self.args.duration
         election_data_dir = os.path.join(
             the_election_config.get("git_rootdir"),
             Globals.get("ROOT_ELECTION_DATA_SUBDIR"),
@@ -309,26 +198,26 @@ mock to a single ballot N times.
             with Shellout.changed_cwd(election_data_dir):
                 Shellout.run(
                     ["git", "pull"],
-                    self.parsed_args.printonly,
-                    self.parsed_args.verbosity,
+                    self.args.printonly,
+                    self.args.verbosity,
                     no_touch_stds=True,
                     timeout=None,
                     check=True,
                 )
-            if self.parsed_args.flush_mode == 2:
+            if self.args.flush_mode == 2:
                 merge_contests = MergeContestsOperation(
                     {
                         "remote": True,
                         "flush": True,
-                        "verbosity": self.parsed_args.verbosity,
-                        "printonly": self.parsed_args.printonly,
+                        "verbosity": self.args.verbosity,
+                        "printonly": self.args.printonly,
                     }
                 )
                 merge_contests.run()
                 tally_contests = TallyContestsOperation(
                     {
-                        "verbosity": self.parsed_args.verbosity,
-                        "printonly": self.parsed_args.printonly,
+                        "verbosity": self.args.verbosity,
+                        "printonly": self.args.printonly,
                     }
                 )
                 tally_contests.run()
@@ -336,9 +225,9 @@ mock to a single ballot N times.
             merge_contests = MergeContestsOperation(
                 {
                     "remote": True,
-                    "minimum_cast_cache": self.parsed_args.minimum_cast_cache,
-                    "verbosity": self.parsed_args.verbosity,
-                    "printonly": self.parsed_args.printonly,
+                    "minimum_cast_cache": self.args.minimum_cast_cache,
+                    "verbosity": self.args.verbosity,
+                    "printonly": self.args.printonly,
                 }
             )
             merge_contests.run()
@@ -347,29 +236,26 @@ mock to a single ballot N times.
             elapsed_time = time.time() - start_time
             if elapsed_time > seconds:
                 break
-        if self.parsed_args.flush_mode in [1, 2]:
+        if self.args.flush_mode in [1, 2]:
             print("Cleaning up remaining unmerged ballots")
             merge_contests = MergeContestsOperation(
                 {
                     "remote": True,
                     "flush": True,
-                    "verbosity": self.parsed_args.verbosity,
-                    "printonly": self.parsed_args.printonly,
+                    "verbosity": self.args.verbosity,
+                    "printonly": self.args.printonly,
                 }
             )
             merge_contests.run()
         # tally the contests
         tally_contests = TallyContestsOperation(
             {
-                "verbosity": self.parsed_args.verbosity,
-                "printonly": self.parsed_args.printonly,
+                "verbosity": self.args.verbosity,
+                "printonly": self.args.printonly,
             }
         )
         tally_contests.run()
 
-    ################
-    # main
-    ################
     # pylint: disable=duplicate-code
     def run(self):
         """Main function - see -h for more info
@@ -390,20 +276,20 @@ mock to a single ballot N times.
         """
 
         # Configure logging
-        Common.configure_logging(self.parsed_args.verbosity)
+        Common.configure_logging(self.args.verbosity)
 
         # Create a VTP ElectionData object if one does not already exist
         the_election_config = ElectionConfig.configure_election()
 
         # If an address was used, use that
         if (
-            self.parsed_args.address
-            or self.parsed_args.state
-            or self.parsed_args.town
-            or self.parsed_args.substreet
+            self.args.address
+            or self.args.state
+            or self.args.town
+            or self.args.substreet
         ):
             the_address = Address.create_address_from_args(
-                self.parsed_args,
+                self.args,
                 [
                     "blank_ballot",
                     "device",
@@ -419,16 +305,11 @@ mock to a single ballot N times.
             blank_ballot = the_election_config.gen_blank_ballot_location(
                 the_address.active_ggos, the_address.ballot_subdir
             )
-        elif self.parsed_args.blank_ballot:
-            blank_ballot = self.parsed_args.blank_ballot
+        elif self.args.blank_ballot:
+            blank_ballot = self.args.blank_ballot
 
         # the VTP scanner mock simulation
-        if self.parsed_args.device in ["scanner", "both"]:
+        if self.args.device in ["scanner", "both"]:
             self.scanner_mockup(the_election_config, blank_ballot)
         else:
             self.server_mockup(the_election_config)
-
-    # End Of Class
-
-
-# EOF
