+++ src/vtp/ops/show_contests_operation.py
@@ -23,7 +23,6 @@ See 'show_contest.py -h' for usage information.
 """
 
 # Standard imports
-import argparse
 import logging
 import os
 import re
@@ -40,51 +39,9 @@ class ShowContestsOperation:
     description (immediately below this) in the source file.
     """
 
-    @staticmethod
-    def parse_arguments(argv):
-        """Parse arguments from a command line or from the constructor"""
-
-        safe_args = Common.cast_thing_to_list(argv)
-        parser = argparse.ArgumentParser(
-            formatter_class=argparse.RawDescriptionHelpFormatter,
-            description="""
-    will print the CVRs (Cast Vote Records) for the supplied contest(s)
-    """,
-        )
-
-        parser.add_argument(
-            "-c",
-            "--contest-check",
-            help="a comma separate list of contests digests to validate/display",
-        )
-        parser.add_argument(
-            "-v",
-            "--verbosity",
-            type=int,
-            default=3,
-            help="0 critical, 1 error, 2 warning, 3 info, 4 debug (def=3)",
-        )
-        parser.add_argument(
-            "-n",
-            "--printonly",
-            action="store_true",
-            help="will printonly and not write to disk (def=True)",
-        )
-        parsed_args = parser.parse_args(safe_args)
-
-        # Validate required args
-        if not parsed_args.contest_check:
-            raise ValueError("The contest check is required")
-        if not bool(re.match("^[0-9a-f,]", parsed_args.contest_check)):
-            raise ValueError(
-                "The contest_check parameter only accepts a comma separated (no spaces) "
-                "list of contest checks/digests to track."
-            )
-        return parsed_args
-
-    def __init__(self, unparsed_args):
+    def __init__(self, args):
         """Only to module-ize the scripts and keep things simple and idiomatic."""
-        self.parsed_args = ShowContestsOperation.parse_arguments(unparsed_args)
+        self.args = args
 
     def validate_digests(self, digests, election_data_dir, error_digests):
         """Will scan the supplied digests for validity.  Will print and
@@ -101,7 +58,7 @@ class ShowContestsOperation:
                         "--batch-check=%(objectname) %(objecttype)",
                         "--buffer",
                     ],
-                    verbosity=self.parsed_args.verbosity,
+                    verbosity=self.args.verbosity,
                     input=input_data,
                     text=True,
                     check=True,
@@ -128,15 +85,12 @@ class ShowContestsOperation:
         if errors:
             raise ValueError(f"Found {errors} invalid digest(s)")
 
-    ################
-    # main
-    ################
     # pylint: disable=duplicate-code
     def run(self):
         """Main function - see -h for more info"""
 
         # Configure logging
-        Common.configure_logging(self.parsed_args.verbosity)
+        Common.configure_logging(self.args.verbosity)
 
         # Create a VTP ElectionData object if one does not already exist
         the_election_config = ElectionConfig.configure_election()
@@ -149,12 +103,10 @@ class ShowContestsOperation:
 
         # First validate the digests
         error_digests = set()
-        self.validate_digests(
-            self.parsed_args.contest_check, election_data_dir, error_digests
-        )
+        self.validate_digests(self.args.contest_check, election_data_dir, error_digests)
         valid_digests = [
             digest
-            for digest in self.parsed_args.contest_check.split(",")
+            for digest in self.args.contest_check.split(",")
             if digest not in error_digests
         ]
         # show/log the digests
@@ -164,20 +116,15 @@ class ShowContestsOperation:
 
 # For future reference just in case . . .
 # this is a loop of shell commands
-#        for digest in self.parsed_args.contest_check.split(','):
+#        for digest in self.args.contest_check.split(','):
 #            if digest not in error_digests:
 #                Shellout.run(['git', 'log', '-1', digest], check=True)
 
 # this does not work well enough either
-#        input_data = '\n'.join(self.parsed_args.contest_check.split(',')) + '\n'
+#        input_data = '\n'.join(self.args.contest_check.split(',')) + '\n'
 #        Shellout.run(
 #            ['git', 'cat-file', '--batch=%(objectname)'],
 #            input=input_data,
 #            text=True,
 #            check=True,
-#            verbosity=self.parsed_args.verbosity)
-
-# End Of Class
-
-
-# EOF
+#            verbosity=self.args.verbosity)
