+++ src/vtp/ops/verify_ballot_receipt_operation.py
@@ -25,7 +25,6 @@ See 'verify_ballot_receipt.py -h' for usage information.
 """
 
 # Standard imports
-import argparse
 import json
 import logging
 import os
@@ -45,75 +44,9 @@ class VerifyBallotReceiptOperation:
     description (immediately below this) in the source file.
     """
 
-    @staticmethod
-    def parse_arguments(argv):
-        """Parse arguments from a command line or from the constructor"""
-
-        safe_args = Common.cast_thing_to_list(argv)
-        parser = argparse.ArgumentParser(
-            formatter_class=argparse.RawDescriptionHelpFormatter,
-            description="""
-Will read a voter's ballot receipt and validate all the digests
-contained therein.  If a contest has been merged to the main branch,
-will report the current ballot tally number (which ballot in the
-actula tally cound is the voter's).
-
-An address is also supported as an argument in which case the last
-ballot check is read from the default location for the specified
-address.
-
-Can also optionally print the ballot's CVRs when a specific ballot
-check row is provided.
-""",
-        )
-
-        Address.add_address_args(parser, True)
-        parser.add_argument(
-            "-f",
-            "--receipt_file",
-            default="",
-            help="specify the ballot receipt location - overrides an address",
-        )
-        parser.add_argument(
-            "-r",
-            "--row",
-            default="",
-            help="specify a row to inspect that row (the first row is 1, not 0)",
-        )
-        parser.add_argument(
-            "-c",
-            "--cvr",
-            action="store_true",
-            help="display the contents of the content CVRs specifying a row",
-        )
-        parser.add_argument(
-            "-x",
-            "--do_not_pull",
-            action="store_true",
-            help="Before tallying the votes, pull the ElectionData repo",
-        )
-        parser.add_argument(
-            "-v",
-            "--verbosity",
-            type=int,
-            default=3,
-            help="0 critical, 1 error, 2 warning, 3 info, 4 debug (def=3)",
-        )
-        #    parser.add_argument("-n", "--printonly", action="store_true",
-        #                            help="will printonly and not write to disk (def=True)")
-
-        parsed_args = parser.parse_args(safe_args)
-
-        # Validate required args
-        if not (parsed_args.receipt_file or (parsed_args.state and parsed_args.town)):
-            raise ValueError(
-                "Either an explicit or implicit (via an address) receipt file must be provided"
-            )
-        return parsed_args
-
-    def __init__(self, unparsed_args):
+    def __init__(self, args):
         """Only to module-ize the scripts and keep things simple and idiomatic."""
-        self.parsed_args = VerifyBallotReceiptOperation.parse_arguments(unparsed_args)
+        self.args = args
 
     # pylint: disable=too-many-arguments   # self is not technically an arg kind-of
     def validate_ballot_lines(self, lines, headers, uids, e_config, error_digests):
@@ -139,7 +72,7 @@ check row is provided.
                     input=input_data,
                     text=True,
                     check=True,
-                    verbosity=self.parsed_args.verbosity,
+                    verbosity=self.args.verbosity,
                     capture_output=True,
                 )
                 .stdout.strip()
@@ -196,7 +129,7 @@ check row is provided.
                     ["git", "log", "--no-walk", "--pretty=format:%H%B"] + row,
                     e_config,
                     grouped_by_uid=False,
-                    verbosity=self.parsed_args.verbosity - 1,
+                    verbosity=self.args.verbosity - 1,
                 )
             elif len(legit_row) > 0:
                 # Only some are legitimate
@@ -204,12 +137,12 @@ check row is provided.
                     ["git", "log", "--no-walk", "--pretty=format:%H%B"] + legit_row,
                     e_config,
                     grouped_by_uid=False,
-                    verbosity=self.parsed_args.verbosity - 1,
+                    verbosity=self.args.verbosity - 1,
                 )
             else:
                 # skip the row - it has no legitimate digests
                 continue
-            if self.parsed_args.row != "" and int(self.parsed_args.row) - 1 == index:
+            if self.args.row != "" and int(self.args.row) - 1 == index:
                 requested_row = cvrs
                 requested_digests = row
             column = -1
@@ -283,7 +216,7 @@ check row is provided.
             contest_batches = Shellout.cvr_parse_git_log_output(
                 ["git", "log", "--topo-order", "--no-merges", "--pretty=format:%H%B"],
                 e_config,
-                verbosity=self.parsed_args.verbosity - 1,
+                verbosity=self.args.verbosity - 1,
             )
             unmerged_uids = {}
             for u_count, uid in enumerate(uids):
@@ -311,21 +244,21 @@ check row is provided.
         # If a row is specified, will print the context index in the
         # actual contest tally - which basically tells the voter 'your
         # contest is in the tally at index N'
-        if self.parsed_args.row:
+        if self.args.row:
             valid_digests = []
-            for digest in lines[int(self.parsed_args.row) - 1]:
+            for digest in lines[int(self.args.row) - 1]:
                 if digest in error_digests:
                     logging.error(
                         "[ERROR]: cannot print CVR for %s (row %s) - it is invalid",
                         digest,
-                        self.parsed_args.row,
+                        self.args.row,
                     )
                     continue
                 valid_digests.append(digest)
                 logging.debug(
                     "%s", json.dumps(requested_row[digest], indent=5, sort_keys=True)
                 )
-            if self.parsed_args.cvr:
+            if self.args.cvr:
                 # Show the CVRs of the row
                 election_data_dir = os.path.join(
                     e_config.get("git_rootdir"),
@@ -351,15 +284,12 @@ check row is provided.
                 "[GOOD]: ballot receipt VALID - no digest errors found\n############"
             )
 
-    ################
-    # main
-    ################
     # pylint: disable=duplicate-code
     def run(self):
         """Main function - see -h for more info"""
 
         # Configure logging
-        Common.configure_logging(self.parsed_args.verbosity)
+        Common.configure_logging(self.args.verbosity)
 
         # Create a VTP ElectionData object if one does not already exist
         the_election_config = ElectionConfig.configure_election()
@@ -368,20 +298,16 @@ check row is provided.
         # remote CVRs branches
         a_ballot = Ballot()
         with Shellout.changed_cwd(a_ballot.get_cvr_parent_dir(the_election_config)):
-            Shellout.run(
-                ["git", "pull"], verbosity=self.parsed_args.verbosity, check=True
-            )
+            Shellout.run(["git", "pull"], verbosity=self.args.verbosity, check=True)
 
         #    import pdb; pdb.set_trace()
-        if self.parsed_args.receipt_file:
+        if self.args.receipt_file:
             # Can read the receipt file directly without any Ballot info
-            self.verify_ballot_receipt(
-                self.parsed_args.receipt_file, the_election_config
-            )
+            self.verify_ballot_receipt(self.args.receipt_file, the_election_config)
         else:
             # Need to use the address to locate the last created receipt file
             the_address = Address.create_address_from_args(
-                self.parsed_args,
+                self.args,
                 ["do_not_pull", "verbosity", "receipt_file", "row", "cvr"],
                 generic_address=True,
             )
@@ -390,8 +316,3 @@ check row is provided.
                 the_election_config, the_address.get("ballot_subdir")
             )
             self.verify_ballot_receipt(receipt_file, the_election_config)
-
-    # End Of Class
-
-
-# EOF
