+++ src/vtp/ops/vote_operation.py
@@ -24,10 +24,6 @@ vote - it simply wraps a call to cast_ballot.py and accept_ballot.py.
 See 'vote.py -h' for usage information.
 """
 
-# Standard imports
-import argparse
-
-from vtp.core.address import Address
 from vtp.core.ballot import Ballot
 from vtp.core.common import Common, Shellout
 from vtp.core.election_config import ElectionConfig
@@ -45,59 +41,15 @@ class VoteOperation:
     description (immediately below this) in the source file.
     """
 
-    @staticmethod
-    def parse_arguments(argv):
-        """Parse arguments from a command line or from the constructor"""
-
-        safe_args = Common.cast_thing_to_list(argv)
-        parser = argparse.ArgumentParser(
-            formatter_class=argparse.RawDescriptionHelpFormatter,
-            description="""
-    Will interactively allow a voter to vote.  Internally it first calls
-    cast_balloy.py followed by accept_ballot.py.  If a specific election
-    address or a specific blank ballot is not specified, a random blank
-    ballot is chosen.
-    """,
-        )
-
-        Address.add_address_args(parser)
-        parser.add_argument(
-            "-m",
-            "--merge_contests",
-            action="store_true",
-            help="Will immediately merge the ballot contests (to main)",
-        )
-        parser.add_argument(
-            "--blank_ballot",
-            help="overrides an address - specifies the specific blank ballot",
-        )
-        parser.add_argument(
-            "-v",
-            "--verbosity",
-            type=int,
-            default=3,
-            help="0 critical, 1 error, 2 warning, 3 info, 4 debug (def=3)",
-        )
-        parser.add_argument(
-            "-n",
-            "--printonly",
-            action="store_true",
-            help="will printonly and not write to disk (def=True)",
-        )
-        return parser.parse_args(safe_args)
-
-    def __init__(self, unparsed_args):
+    def __init__(self, args):
         """Only to module-ize the scripts and keep things simple and idiomatic."""
-        self.parsed_args = VoteOperation.parse_arguments(unparsed_args)
+        self.args = args
 
-    ################
-    # main
-    ################
     def run(self):
         """Main function - see -h for more info"""
 
         # Configure logging
-        Common.configure_logging(self.parsed_args.verbosity)
+        Common.configure_logging(self.args.verbosity)
 
         # Create a VTP ElectionData object if one does not already exist
         the_election_config = ElectionConfig.configure_election()
@@ -108,34 +60,34 @@ class VoteOperation:
         with Shellout.changed_cwd(a_ballot.get_cvr_parent_dir(the_election_config)):
             Shellout.run(
                 ["git", "pull"],
-                printonly=self.parsed_args.printonly,
-                verbosity=self.parsed_args.verbosity,
+                printonly=self.args.printonly,
+                verbosity=self.args.verbosity,
                 check=True,
             )
 
         # If an address was used, use that
         cast_args = {
-            "verbosity": self.parsed_args.verbosity,
-            "printonly": self.parsed_args.printonly,
+            "verbosity": self.args.verbosity,
+            "printonly": self.args.printonly,
         }
         accept_args = {
-            "verbosity": self.parsed_args.verbosity,
-            "printonly": self.parsed_args.printonly,
+            "verbosity": self.args.verbosity,
+            "printonly": self.args.printonly,
         }
-        if not self.parsed_args.blank_ballot:
-            if self.parsed_args.state:
-                cast_args["state"] = self.parsed_args.state
-                accept_args["state"] = self.parsed_args.state
-            if self.parsed_args.town:
-                cast_args["town"] = self.parsed_args.town
-                accept_args["town"] = self.parsed_args.town
-            if self.parsed_args.substreet:
-                cast_args["substreet"] = self.parsed_args.substreet
-            if self.parsed_args.address:
-                cast_args["address"] = self.parsed_args.address
+        if not self.args.blank_ballot:
+            if self.args.state:
+                cast_args["state"] = self.args.state
+                accept_args["state"] = self.args.state
+            if self.args.town:
+                cast_args["town"] = self.args.town
+                accept_args["town"] = self.args.town
+            if self.args.substreet:
+                cast_args["substreet"] = self.args.substreet
+            if self.args.address:
+                cast_args["address"] = self.args.address
         else:
-            cast_args["blank_ballot"] = self.parsed_args.blank_ballot
-            accept_args["blank_ballot"] = self.parsed_args.blank_ballot
+            cast_args["blank_ballot"] = self.args.blank_ballot
+            accept_args["blank_ballot"] = self.args.blank_ballot
 
         # Basically only do as little as necessary to call cast_ballot.py
         # followed by accept_ballot.py
@@ -145,8 +97,3 @@ class VoteOperation:
         # Accept the ballot
         a_accept_ballot_operation = AcceptBallotOperation(accept_args)
         a_accept_ballot_operation.run()
-
-    # End Of Class
-
-
-# EOF
